Arreglando permisos.

+++ backend/main/resources/usuario.py
@@ -5,6 +5,7 @@ from main.models import UsuarioModel, PrestamoModel, ReseñaModel
 from sqlalchemy import func, desc, asc
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from main.auth.decorators import role_required
+from werkzeug.security import generate_password_hash
 
 class Usuarios(Resource):
     @role_required(roles = ["Admin"])
@@ -92,30 +93,50 @@ class Usuarios(Resource):
                   'page': page
                 })
         
-    @role_required(roles = ["Admin"])
+    @role_required(roles = ["Admin", "Bibliotecario"])
     def post(self):
         usuario = UsuarioModel.from_json(request.get_json())
-        try:
-            db.session.add(usuario)
-            db.session.commit()
-        except:
-            return "Formato de datos incorrecto.", 400
-        return usuario.to_json(), 201
+        
+        jwt_identity = get_jwt_identity()
+        current_user = db.session.query(UsuarioModel).get_or_404(jwt_identity)
+        if current_user.rol == "Bibliotecario" and (usuario.rol == "Admin" or usuario.rol == "Bibliotecario"):
+            return "Permiso denegado.", 403
+        else:
+            try:
+                db.session.add(usuario)
+                db.session.commit()
+            except:
+                return "Formato de datos incorrecto.", 400
+            return usuario.to_json(), 201
 
 class Usuario(Resource):
-    @jwt_required(optional=True)
+    @jwt_required()
     def get(self, id): 
         try:
             usuario = db.session.query(UsuarioModel).get_or_404(id)
         except:
             return "ID inexistente.", 404
-        
-        if get_jwt_identity():
-            current_identity = get_jwt_identity()
-            if int(current_identity) == int(id):
+    
+        jwt_identity = get_jwt_identity()
+        current_user = db.session.query(UsuarioModel).get_or_404(jwt_identity)
+        if current_user.rol == "Usuario": 
+            if usuario.rol == "Usuario": 
+                if int(jwt_identity) != int(id):
+                    return usuario.to_json_short()
+                else:
+                    return usuario.to_json_complete()
+            else:
+                return "Permiso denegado.", 403
+
+        elif current_user.rol == "Bibliotecario":
+            if usuario.rol == "Admin":
+                return "Permiso denegado.", 403
+            else:
                 return usuario.to_json_complete()
-        return usuario.to_json()
-        
+
+        else:
+            return usuario.to_json_complete()
+    
     @jwt_required()
     def put(self, id):
         try:
@@ -123,18 +144,37 @@ class Usuario(Resource):
         except:
             return "ID inexistente.", 404
         data = request.get_json().items()
+        jwt_identity = get_jwt_identity()
+        current_user = db.session.query(UsuarioModel).get_or_404(jwt_identity)
+        
         for key, value in data:
+            if key == "id" or key == "rol" or key == "reseñas" or key == "prestamos" or key == "notificaciones":
+                continue
+            if key == "contraseña":
+                value = generate_password_hash(value)
             setattr(usuario, key, value)
-        db.session.add(usuario)
-        db.session.commit()
-        return usuario.to_json() , 201    
-    
-    @role_required(roles = ["Admin","Bibliotecario","Usuario"])
+        
+        if (current_user.rol == "Usuario" or current_user.rol == "Bibliotecario") and int(jwt_identity) != int(id):
+            return "Permiso denegado.", 403
+        else:
+            db.session.add(usuario)
+            db.session.commit()
+            return usuario.to_json(), 201    
+
+    @jwt_required()    
     def delete(self, id):
         try:
             usuario = db.session.query(UsuarioModel).get_or_404(id)
         except:
             return "ID inexistente.", 404
-        db.session.delete(usuario)
-        db.session.commit()
-        return usuario.to_json(), 204
+
+        jwt_identity = get_jwt_identity()
+        current_user = db.session.query(UsuarioModel).get_or_404(jwt_identity)
+        if current_user.rol == "Usuario" and int(jwt_identity) != int(id):
+            return "Permiso denegado.", 403
+        if current_user.rol == "Bibliotecario" and (usuario.rol == "Admin" or usuario.rol == "Bibliotecario"):
+            return "Permiso denegado.", 403
+        else:
+            db.session.delete(usuario)
+            db.session.commit()
+            return usuario.to_json(), 204
